Prepare `v7.0.0` Release
ADDED:

  * TER-391: New `-l`/`--pipe-plan` arguments to pipe the `plan` output
to the command of your choice. That command can be specified with the
new `--pipe-plan-command` argument or configured in the `config.yml`
file as `pipe_plan_command: "my custom command"`.
  * TER-396: Allow mutiple states backends. We can now choose per stack
which state backend to use

UPDATES:

  * TER-46: Bump Python dependencies
  * TER-364: Upgrade `azure cli` pip package to .64
  * TER-393: Support terraform release candidates
  * TER-395: Prioritize bootstrap template parameter, allow to bootstrap
a non cloud template
  * TER-404: Make `init` optional via config when triggering `plan` or
`apply`
  * TER-406: Update README documentation

BREAKING:

  * TER-388: Improve `bootstrap` behavior, no longer trigger `terraform
init` automatically (old behavior is available via TER-404, can be set
in config)